backend/ml/requirements.txt
tensorflow==2.14.0
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
numpy==1.24.4
pydantic==2.5.0
python-multipart==0.0.6
scikit-learn==1.3.2
matplotlib==3.8.2
requests==2.31.0
